moved reinforce-syn to right spot
