+++ ngclearn/components/synapses/hebbian/__init__.py
@@ -3,3 +3,4 @@ from .traceSTDPSynapse import TraceSTDPSynapse
 from .expSTDPSynapse import ExpSTDPSynapse
 from .eventSTDPSynapse import EventSTDPSynapse
 from .BCMSynapse import BCMSynapse
+
