+++ ngclearn/components/synapses/modulated/__init__.py
@@ -1,1 +1,3 @@
 from .MSTDPETSynapse import MSTDPETSynapse
+from .REINFORCESynapse import REINFORCESynapse
+
